requirements.txt
ffmpeg-python>=0.2.0
boto3>=1.28.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0
aiofiles>=23.2.0
